refactor(reader): split data_access.rs by read path (#1122)
Behavior-preserving module split of the ~3,988-line
reader/data_access.rs into data_access/{mod,model,bti,sequential,
compaction}.rs per epic #1116, by read path:

- mod.rs: shared entry points + infra (get, build_v5_parser,
  new_scan_cursor, read_next_block, read_value_at_offset,
  get_cached_data, stitch_all_chunks + stitch_and_parse_*,
  prepare_delta_scan) + pub-use re-exports (ClusteringSlice,
  table_ids_match) so the public surface and sibling
  scan_stream_windowed use-path are unchanged.
- model.rs: path-agnostic types/helpers (ClusteringSlice,
  ClusteringRowWindow, BtiLookupStep, physical_byte_bounds_for_slice,
  table_ids_match[_strict], bti_lookup_step, sort_by_token_order*,
  SCAN_FOR_KEY_CALLS) + their unit tests.
- bti.rs: BTI trie point lookup / single-partition seek / clustering
  window / whole-Data.db scan + BTI math tests.
- sequential.rs: scan, get_all_entries, scan_stream, scan_for_key,
  sequential_scan, scan_with_cell_metadata + integration tests.
- compaction.rs: timestamp-preserving iteration, distinct partition
  keys, bounded streaming compaction driver.

No semantic changes. Tests relocated with their code (count
preserved). Visibility widened only where a cross-submodule reach
required it (pub(in ...reader)).

3237 passed / 0 failed; clippy -D warnings clean (default +
all-features); minimal-features build green.
