+++ cqlite-core/src/storage/sstable/reader/data_access/mod.rs
@@ -1,0 +1,505 @@
+//! Data access methods for SSTableReader.
+//!
+//! This module contains all methods related to reading data from SSTables,
+//! split by read path:
+//! - [`model`] — shared model types and free helpers (clustering slice,
+//!   table-id matching, token-order sorting, the `scan_for_key` counter).
+//! - [`bti`] — BTI ("da") trie-resolved point lookups, single-partition seeks,
+//!   clustering-slice narrowing, and the whole-Data.db BTI scan.
+//! - [`sequential`] — sequential / index-driven range and full scans, the
+//!   `scan_for_key` fallback, cell-metadata scan, and the streaming scan.
+//! - [`compaction`] — timestamp-preserving compaction iteration and the bounded
+//!   streaming compaction driver.
+//!
+//! `mod.rs` keeps the shared entry points and the I/O / parser infrastructure
+//! every read path builds on (`get`, `build_v5_parser`, `new_scan_cursor`,
+//! `read_next_block`, the offset-read helpers, the chunk-stitch helpers, and the
+//! delta-scan bridge), plus the public-surface re-exports.
+
+mod bti;
+mod compaction;
+mod model;
+mod sequential;
+
+// Public surface re-export (unchanged: `reader::mod` re-exports
+// `data_access::ClusteringSlice`).
+pub use model::ClusteringSlice;
+
+// Re-export for the sibling `scan_stream_windowed` module, which references
+// `data_access::table_ids_match` (unchanged path).
+pub(in crate::storage::sstable::reader) use model::table_ids_match;
+
+use super::source::ScanCursor;
+use super::SSTableReader;
+use crate::parser::DataFormat;
+use crate::types::{CellWriteMetadata, TableId, Value};
+use crate::{Error, Result, RowKey};
+use log::{debug, warn};
+use std::io::SeekFrom;
+use tokio::io::AsyncSeekExt;
+
+impl SSTableReader {
+    /// Return `true` when Data.db uses the V5CompressedLegacy NB chunked format and
+    /// therefore requires all chunks to be stitched before parsing.
+    ///
+    /// The correct predicate is:
+    ///   data_format == V5CompressedLegacy  AND  is_nb_format()
+    ///
+    /// Rationale:
+    /// - `V5CompressedLegacy` identifies the row serialization format (u16 length
+    ///   prefixes, legacy encoding) used by all Cassandra 5 'nb' SSTables.
+    /// - `is_nb_format()` identifies the chunked-compression read path. It intentionally
+    ///   EXCLUDES `V5_0Uncompressed`, which uses the same row format but stores data as
+    ///   a single contiguous block (no chunk boundaries, no stitching needed).
+    /// - Using `is_compressed` (compression_reader.is_some()) would be wrong for NB
+    ///   format because the per-chunk decompression is handled inside `stitch_and_parse_all_chunks`,
+    ///   and `is_compressed` may differ from `is_nb_format` for edge-case versions.
+    pub(super) fn requires_chunk_stitching(&self) -> bool {
+        let data_format = self.header.cassandra_version.data_format();
+        matches!(data_format, DataFormat::V5CompressedLegacy)
+            && self.header.cassandra_version.is_nb_format()
+    }
+
+    /// Get a value by key from the SSTable
+    pub async fn get(&self, table_id: &TableId, key: &RowKey) -> Result<Option<Value>> {
+        use crate::observability::{self as obs, catalog};
+
+        // Issue #831 / #909: BTI ("da") readers resolve partitions via the
+        // Partitions.db trie (O(log n)), never via Index.db (absent for BTI) or
+        // the sequential scan. The trie is the AUTHORITATIVE presence oracle for a
+        // BTI SSTable — it answers present/absent definitively — so we branch here
+        // BEFORE the bloom-filter pre-check. Skipping the bloom filter for BTI is
+        // both correct (the trie is authoritative; bloom is only an optimization)
+        // and necessary: a writer-produced Filter.db whose hashing does not match
+        // the reader's would otherwise cause false negatives and drop live
+        // partitions (the writer→reader roundtrip #909 must read back). It also
+        // guarantees a BTI get() can never fall through to scan_for_key.
+        if self.bti_partitions_db.is_some() {
+            return self.bti_point_lookup(table_id, key).await;
+        }
+
+        // First check bloom filter if available
+        if let Some(bloom_filter) = &self.bloom_filter {
+            let present = bloom_filter.might_contain(key.as_bytes());
+            obs::add_counter(
+                catalog::READ_BLOOM_CHECKS,
+                1,
+                &[
+                    (
+                        catalog::attr::RESULT,
+                        if present { "hit" } else { "miss" }.into(),
+                    ),
+                    (
+                        catalog::attr::SSTABLE_FORMAT,
+                        self.sstable_format_label().into(),
+                    ),
+                ],
+            );
+            if !present {
+                return Ok(None);
+            }
+        }
+
+        // Use index for efficient lookup if available
+        if let Some(index) = &self.index {
+            if let Some(entry) = index.find_entry(table_id, key).await? {
+                // When Index.db reports size=0 (Cassandra 5.0), fall back to sequential scan
+                if entry.size == 0 {
+                    log::debug!(
+                        "Index reports size=0 for key {:?}, using sequential scan fallback",
+                        key
+                    );
+                    return self.scan_for_key(table_id, key).await;
+                }
+
+                // Index offsets are relative to data section start - adjust for header
+                let file_offset = entry.offset + self.actual_header_size as u64;
+                return self.read_value_at_offset(file_offset, entry.size).await;
+            }
+
+            // Issue #517: The SSTableIndex is built from Index.db key *digests* (16-byte
+            // Murmur3 hashes), not raw partition key bytes.  A raw-key lookup via
+            // find_entry() always misses.  Fall back to scan_for_key() so that get()
+            // and scan() agree on which partitions exist.
+            log::debug!(
+                "Index lookup returned no entry for key {:?} (possible digest/raw-key mismatch), \
+                 falling back to sequential scan",
+                key
+            );
+            return self.scan_for_key(table_id, key).await;
+        } else {
+            // No index at all — fall back to sequential scan
+            return self.scan_for_key(table_id, key).await;
+        }
+    }
+
+    /// Stitch all compressed chunks and parse as a single buffer (V5CompressedLegacy)
+    ///
+    /// This helper method extracts the stitching logic from get_all_entries so it can be
+    /// reused by sequential_scan and other methods that need to handle V5CompressedLegacy
+    /// format where partitions can span chunk boundaries.
+    pub(super) async fn stitch_and_parse_all_chunks(
+        &self,
+        cursor: &ScanCursor,
+        schema: Option<&crate::schema::TableSchema>,
+    ) -> Result<Vec<(TableId, RowKey, Value)>> {
+        let stitched_buffer = self.stitch_all_chunks(cursor).await?;
+        let parser = self.build_v5_parser();
+
+        // Get schema (use provided schema or reader's schema)
+        let reader_schema;
+        let table_schema = if let Some(s) = schema {
+            Some(s)
+        } else {
+            reader_schema = self.get_table_schema(None);
+            reader_schema.as_ref()
+        };
+
+        // Parse the stitched decompressed buffer
+        let entries = parser.parse_block(&stitched_buffer, table_schema, self)?;
+        log::debug!(
+            "stitch_and_parse_all_chunks: Parsed {} entries from stitched buffer",
+            entries.len()
+        );
+
+        Ok(entries)
+    }
+
+    /// Like [`stitch_and_parse_all_chunks`] but also returns per-cell write metadata.
+    ///
+    /// Used when `ProjectionFlags::include_cell_metadata` is set (issue #693).
+    ///
+    /// [`stitch_and_parse_all_chunks`]: Self::stitch_and_parse_all_chunks
+    pub(super) async fn stitch_and_parse_all_chunks_with_metadata(
+        &self,
+        cursor: &ScanCursor,
+        schema: Option<&crate::schema::TableSchema>,
+    ) -> Result<
+        Vec<(
+            TableId,
+            RowKey,
+            Value,
+            std::collections::HashMap<String, CellWriteMetadata>,
+        )>,
+    > {
+        let stitched_buffer = self.stitch_all_chunks(cursor).await?;
+        let parser = self.build_v5_parser();
+
+        let reader_schema;
+        let table_schema = if let Some(s) = schema {
+            Some(s)
+        } else {
+            reader_schema = self.get_table_schema(None);
+            reader_schema.as_ref()
+        };
+
+        let entries =
+            parser.parse_block_with_cell_metadata(&stitched_buffer, table_schema, self)?;
+        log::debug!(
+            "stitch_and_parse_all_chunks_with_metadata: Parsed {} entries with metadata",
+            entries.len()
+        );
+
+        Ok(entries)
+    }
+
+    /// Read, decompress, and concatenate every compressed chunk of the data
+    /// section into a single buffer.
+    ///
+    /// V5CompressedLegacy partitions can span chunk boundaries, so the whole
+    /// data section must be stitched before parsing. The returned buffer is
+    /// bounded by the *uncompressed data-section size* — it scales with on-disk
+    /// bytes, not row count (issue #790).
+    ///
+    /// Precondition: the caller has seeked `cursor`'s file to the start of the
+    /// data section (the cursor's chunk index starts at 0 when freshly minted).
+    pub(super) async fn stitch_all_chunks(&self, cursor: &ScanCursor) -> Result<Vec<u8>> {
+        use crate::storage::sstable::compression::Compression;
+
+        // Pre-allocate buffer for ~2.5MB (estimated max size for test data)
+        let mut stitched_buffer = Vec::with_capacity(2_500_000);
+
+        // Incompressible-chunk fallback (Bug #639, epic #970): Cassandra stores a
+        // chunk RAW (not compressed) when its compressed length would meet or
+        // exceed `max_compressed_length`. `ChunkDecompressor::decompress_chunk`
+        // already honours this, but the stitch path did not — it blindly tried to
+        // LZ4/Snappy/etc-decode a raw chunk, which fails on the `incompressible`
+        // fixture. Mirror the writer rule here: when the (CRC-stripped) chunk
+        // length >= max_compressed_length, the bytes are already plaintext.
+        // Authority: CompressedSequentialWriter.java:160-177.
+        let max_compressed_length = self
+            .compression_info
+            .as_ref()
+            .map(|ci| ci.max_compressed_length as usize)
+            .unwrap_or(usize::MAX);
+
+        let mut chunk_count = 0;
+        while let Some(compressed_chunk) = self.read_next_block(cursor).await? {
+            let decompressed_chunk = if compressed_chunk.len() >= max_compressed_length {
+                // Stored uncompressed by Cassandra — pass the raw bytes through.
+                log::debug!(
+                    "stitch_all_chunks: chunk {} is incompressible (len={} >= max_compressed_length={}), using raw bytes",
+                    chunk_count,
+                    compressed_chunk.len(),
+                    max_compressed_length
+                );
+                compressed_chunk
+            } else if let Some(compression_reader) = &self.compression_reader {
+                let compression = Compression::new(*compression_reader.algorithm())?;
+                match compression.decompress(&compressed_chunk) {
+                    Ok(decompressed) => decompressed,
+                    Err(e) => {
+                        return Err(Error::corruption(format!(
+                            "stitch_all_chunks: Failed to decompress chunk {}: {}",
+                            chunk_count, e
+                        )));
+                    }
+                }
+            } else {
+                // No compression (should not happen for V5CompressedLegacy)
+                log::warn!("stitch_all_chunks: No compression reader, using raw chunk data");
+                compressed_chunk
+            };
+
+            stitched_buffer.extend_from_slice(&decompressed_chunk);
+            chunk_count += 1;
+        }
+
+        log::debug!(
+            "stitch_all_chunks: Stitched {} chunks, total buffer: {} bytes",
+            chunk_count,
+            stitched_buffer.len()
+        );
+
+        Ok(stitched_buffer)
+    }
+
+    /// Build a [`V5CompressedLegacyParser`] configured from this reader's header,
+    /// statistics (EncodingStats), version gates, and UDT registry.
+    ///
+    /// [`V5CompressedLegacyParser`]: crate::storage::sstable::reader::parsing::V5CompressedLegacyParser
+    pub(super) fn build_v5_parser(
+        &self,
+    ) -> crate::storage::sstable::reader::parsing::V5CompressedLegacyParser {
+        let keyspace = self.header.keyspace.clone();
+        let table_name = self.header.table_name.clone();
+
+        // Extract EncodingStats from statistics_reader (if available)
+        let (min_timestamp, min_local_deletion_time, min_ttl) =
+            if let Some(stats_reader) = &self.statistics_reader {
+                let ts_stats = &stats_reader.statistics().timestamp_stats;
+                (
+                    ts_stats.min_timestamp,
+                    ts_stats.min_deletion_time,
+                    ts_stats.min_ttl,
+                )
+            } else {
+                (0, 0, None)
+            };
+
+        let parser = crate::storage::sstable::reader::parsing::V5CompressedLegacyParser::new(
+            keyspace,
+            table_name,
+            min_timestamp,
+            min_local_deletion_time,
+            min_ttl,
+        )
+        // VG1: thread VersionGates from SSTableReader down to row parser so
+        // that VG3 can flip gate-sensitive code paths without re-deriving gates.
+        .with_version_gates(self.version_gates.clone());
+        // Add UDT registry if available for UDT-aware collection parsing (Issue #238)
+        if let Some(ref registry) = self.udt_registry {
+            parser.with_udt_registry(registry.clone())
+        } else {
+            parser
+        }
+    }
+
+    /// Read value at a specific offset with caching
+    pub async fn read_value_at_offset(&self, offset: u64, size: u32) -> Result<Option<Value>> {
+        use crate::parser::header::CassandraVersion;
+        use crate::storage::sstable::compression::Compression;
+
+        // Size must be non-zero for offset-based reading
+        if size == 0 {
+            return Err(Error::corruption(format!(
+                "Cannot read value at offset {} with size=0. This should have been caught earlier and handled via sequential scan.",
+                offset
+            )));
+        }
+
+        // Use cached reading with metrics tracking
+        let buffer = self.get_cached_data(offset, size).await?;
+
+        // Decompress if needed
+        let data = if let Some(compression_reader) = &self.compression_reader {
+            let compression = Compression::new(*compression_reader.algorithm())?;
+            match compression.decompress(&buffer) {
+                Ok(decompressed) => {
+                    debug!(
+                        "Successfully decompressed {} bytes to {} bytes",
+                        buffer.len(),
+                        decompressed.len()
+                    );
+                    decompressed
+                }
+                Err(e) => {
+                    // For modern formats (4.x/5.x), decompression failure is an error
+                    if self.header.cassandra_version != CassandraVersion::Legacy {
+                        return Err(Error::corruption(format!(
+                            "Decompression failed for modern format at offset={}, size={}, algorithm={:?}: {}",
+                            offset,
+                            size,
+                            compression_reader.algorithm(),
+                            e
+                        )));
+                    } else {
+                        // Only allow fallback for legacy formats
+                        warn!(
+                            "Decompression failed for legacy format ({}), using raw data",
+                            e
+                        );
+                        debug!(
+                            "First 32 bytes of raw data: {:02x?}",
+                            &buffer[..std::cmp::min(32, buffer.len())]
+                        );
+                        buffer
+                    }
+                }
+            }
+        } else {
+            buffer
+        };
+
+        // TODO: Parse value using schema-driven type information
+        // For now, preserve raw data until schema is available
+        let value = Value::Blob(data.to_vec());
+
+        // Extract write time from value (placeholder - would need to be parsed from SSTable)
+        let _write_time = std::time::SystemTime::now()
+            .duration_since(std::time::UNIX_EPOCH)
+            .map(|d| d.as_micros() as i64)
+            .unwrap_or_else(|e| {
+                warn!("Failed to get system time: {}; using fallback value 0", e);
+                0
+            });
+
+        // Filter out tombstones and expired data
+        if !self.filter_tombstone(&value) {
+            return Ok(None);
+        }
+
+        Ok(Some(value))
+    }
+
+    /// Read block with caching support and hit/miss tracking
+    async fn get_cached_data(&self, block_offset: u64, size: u32) -> Result<Vec<u8>> {
+        use crate::parser::header::CassandraVersion;
+        use crate::storage::sstable::compression::Compression;
+        use tokio::io::AsyncReadExt;
+
+        // Calculate block identifier based on offset and size
+        let _block_id = block_offset;
+
+        // For now, always read from disk and track as cache miss
+        self.record_cache_miss();
+
+        // Read from disk
+        let mut file = self.file.lock().await;
+        file.seek(SeekFrom::Start(block_offset)).await?;
+
+        let mut buffer = vec![0u8; size as usize];
+        file.read_exact(&mut buffer).await?;
+        drop(file); // Release file lock early
+
+        // Decompress if needed
+        let data = if let Some(compression_reader) = &self.compression_reader {
+            let compression = Compression::new(*compression_reader.algorithm())?;
+            match compression.decompress(&buffer) {
+                Ok(decompressed) => decompressed,
+                Err(e) => {
+                    // Handle decompression errors based on format
+                    if self.header.cassandra_version != CassandraVersion::Legacy {
+                        return Err(Error::corruption(format!(
+                            "Decompression failed at offset={}, size={}: {}",
+                            block_offset, size, e
+                        )));
+                    } else {
+                        buffer // Fall back to raw data for legacy formats
+                    }
+                }
+            }
+        } else {
+            buffer
+        };
+
+        Ok(data)
+    }
+
+    /// Mint a fresh, independent cursor for one scan (issue #815).
+    ///
+    /// Each cursor owns a private file handle (or mmap cursor) and chunk index,
+    /// so concurrent scans on this reader never share a mutable file position —
+    /// they run in parallel without the per-scan serialization #805 required.
+    pub(in crate::storage::sstable::reader) async fn new_scan_cursor(&self) -> Result<ScanCursor> {
+        Ok(ScanCursor::new(self.scan_source.open(&self.file_path).await?))
+    }
+
+    /// Read the next block from a scan-local `cursor` (its own file position and
+    /// chunk index). See [`Self::new_scan_cursor`].
+    pub(in crate::storage::sstable::reader) async fn read_next_block(
+        &self,
+        cursor: &ScanCursor,
+    ) -> Result<Option<Vec<u8>>> {
+        use super::block_io;
+        block_io::read_next_block(
+            &cursor.file,
+            &self.header.cassandra_version,
+            &self.config,
+            &self.compression_info,
+            &cursor.chunk_index,
+            self.actual_header_size as u64,
+        )
+        .await
+    }
+
+    /// Prepare for a delta-scan pass: stitch all compressed chunks of the data
+    /// section and return the decompressed buffer together with a pre-configured
+    /// parser.
+    ///
+    /// Uses its own per-scan cursor (issue #815), so it no longer needs the
+    /// caller to serialize against concurrent reads. This method is gated on the
+    /// `delta-scan` feature and is the only bridge between the SSTableReader
+    /// internals and the `delta_scan` module, which cannot access private
+    /// helpers directly.
+    ///
+    /// The `schema` parameter is not used here — it is threaded through the
+    /// caller's `parse_block_emit_delta` invocation instead.  The parser is
+    /// built via `build_v5_parser()` which handles version-gates and UDT
+    /// registry without needing the schema at construction time.
+    #[cfg(feature = "delta-scan")]
+    pub async fn prepare_delta_scan(
+        &self,
+    ) -> Result<(Vec<u8>, super::parsing::V5CompressedLegacyParser)> {
+        use tokio::io::AsyncSeekExt;
+
+        // Seek the per-scan cursor to the start of the data section.
+        let cursor = self.new_scan_cursor().await?;
+        let header_size = self.calculate_header_size();
+        {
+            let mut file_guard = cursor.file.lock().await;
+            file_guard
+                .seek(std::io::SeekFrom::Start(header_size as u64))
+                .await?;
+        }
+
+        // Stitch all compressed chunks (bounded by uncompressed data-section size).
+        let stitched = self.stitch_all_chunks(&cursor).await?;
+
+        // Build a parser (re-using the existing builder so version-gates and
+        // UDT registry are threaded through correctly).
+        let parser = self.build_v5_parser();
+
+        Ok((stitched, parser))
+    }
+}
